moved whatsapp function to util
